Utvider endepunkt for å hente alle dokumenter på sak til å også kunne hente alle dokumenter på bruker + renames/rydding
